refactor: merge storage modules
Merging storage modules will permit to reduce the overall number of Maven modules within the project. This will reduce a bit the project setup complexity and ease writing of integration tests.
