+++ src/patch-place-break/patch-place-break-core/src/test/resources/fr/djaytan/minecraft/jobsreborn/patchplacebreak/core/PatchPlaceBreakCoreIntegrationTest/whenEnabling_withCustomSetup_givenConfigFileTemplate.conf
