+++ src/storage/storage-sql/src/main/java/fr/djaytan/minecraft/jobsreborn/patchplacebreak/storage/sql/serializer/StringSerializer.java
