+++ src/storage/storage-sql/src/main/resources/db/migration/mysql/V1__Initial_version.sql
