+++ src/storage/storage-sql/src/main/resources/db/migration/mysql/V2_1_0__Integer_type_for_coordinates.sql
