+++ src/storage/storage-sql/src/main/resources/db/migration/sqlite/V1__Initial_version.sql
