+++ src/storage/storage-sql/src/main/resources/db/migration/sqlite/V2_1_3__Remove_uuid_field.sql
